refactor: break Piggyback into on input and output
Since for the input case, we always have the "exitTarget" saved
during start in-flight exit, we never need outputType and outputGuardPreimage
as arguments. Also there are some logic branching already, seperate to
two functions instead of one.
